Feature/remove activation endpoints (#43)
* Redefinition of activate/deactivate endpoints

* Fixing WebsiteRecord update

* Adding celery working stat

+++ backend/crawler/tasks/crawler.py
@@ -6,7 +6,7 @@ from redbeat import RedBeatSchedulerEntry
 from api.models import WebsiteRecord
 
 from .transformer import transform_graph, persist_graph
-from crawler.celery import app
+from crawler.celery import app, celery_is_active
 
 
 @app.task(bind=True)
@@ -25,13 +25,15 @@ def schedule_periodic_crawler_task(url: str, regex: str, record_id: int, interva
 
 
 def manage_tasks(record: WebsiteRecord, reschedule: bool = False):
-    if 'test' in sys.argv:
+    if 'test' in sys.argv or not celery_is_active():
         return 0
 
     if not reschedule:
         if record.interval:
             record.job_id = f'redbeat:task:{record.id}'
-            return schedule_periodic_crawler_task(record.url, record.regex, record.id, record.interval).key
+            task_id = schedule_periodic_crawler_task(record.url, record.regex, record.id, record.interval).key
+            record.save()
+            return task_id
         elif not record.interval:
             return run_crawler_task.delay(record.url, record.regex, record.id).id
     else:
@@ -39,16 +41,26 @@ def manage_tasks(record: WebsiteRecord, reschedule: bool = False):
             RedBeatSchedulerEntry.from_key(record.job_id, app=app).delete()
             return schedule_periodic_crawler_task(record.url, record.regex, record.id, record.interval).key
         elif not record.active:
+            entry = RedBeatSchedulerEntry.from_key(record.job_id, app=app)
+            if entry:
+                entry.delete()
+
             return run_crawler_task.delay(record.url, record.regex, record.id).id
 
 
 def stop_periodic_task(record: WebsiteRecord):
+    if not celery_is_active():
+        return
+
     if 'test' not in sys.argv:
         if record.job_id and record.interval:
             RedBeatSchedulerEntry.from_key(record.job_id, app=app).delete()
 
 
 def start_periodic_task(record: WebsiteRecord):
+    if not celery_is_active():
+        return
+
     if 'test' not in sys.argv:
         if not record.job_id and record.interval:
             schedule_periodic_crawler_task(record.url, record.regex, record.id, record.interval)
